+++ src/index.ts
@@ -8,6 +8,7 @@ export * from "./lib/Layout/PanelSideBarLayout/PanelSideBar/Definitions/PanelIte
 
 // Sidebar layout
 export * from "./lib/Layout/SideBarLayout/SideBarLayout";
+export * from "./lib/Layout/SideBarLayout/SideBarLayoutContext";
 export * from "./lib/SideBar/SideBarMenuContext";
 export * from "./lib/Paging/Paging";
 export * from "./lib/SideBar/ISideBarMenuItem";